feat: new clean_images_by_date.py script
